[mac] Phases M7, M8, M11: strip old AppKit chrome + flag, native shell is now the only shell
Deletes the legacy AppKit shell entirely:
- SidebarViewController, FileExplorerView (+ FlatSectionOutlineView,
  ClearlySidebarRowView, ClearlySidebarBackgroundView)
- TabBarView, ClearlySegmentedControl, ClearlyButtonStyle
- ContentView (extracts FocusedValueKeys + FocusedValuesModifier +
  Notification.Name extension into Clearly/Native/NativeShellSupport.swift)
- ClearlySplitViewController, SoftDividerSplitView, DetailView, NoFileView
- UseNativeMacShell flag + NativeMainWindowContent wrapper
- LauncherSceneMarker + repositionTrafficLights (no more manual traffic-light
  repositioning; native titlebar renders correctly in the unified toolbar)

ClearlyAppDelegate keeps its lifecycle/event-monitor/menu-injection role but
loses the imperative createMainWindow() + MainWindowDelegateImpl +
setSidebarVisible helpers. SwiftUI's Window("Clearly", id: "main") scene
owns window creation and close. ⌘L toggles the NavigationSplitView
sidebar via responder chain against NSSplitViewController.toggleSidebar(_:).

Outline mounts on native .inspector() in the detail column; Find and
JumpToLine stay as top overlays; Backlinks is an inline bottom panel;
QuickSwitcher stays a floating NSPanel. Menu commands move onto the
new main Window scene.
