--- Clearly/SidebarViewController.swift
@@ -1,186 +1,0 @@
-import AppKit
-import SwiftUI
-import ClearlyCore
-
-/// Pure AppKit sidebar view controller — no SwiftUI hosting view for the main content.
-/// The outline view is built directly in AppKit so the background is fully controlled.
-@MainActor
-class SidebarViewController: NSViewController {
-    let workspace: WorkspaceManager
-    private var outlineCoordinator: FileExplorerOutlineView.Coordinator?
-    private var scrollView: NSScrollView?
-    private var emptyHostingView: NSHostingView<FileExplorerEmptyView>?
-    private var updateTimer: Timer?
-    private var headerHeightConstraint: NSLayoutConstraint?
-    private var fullscreenObservers: [NSObjectProtocol] = []
-
-    init(workspace: WorkspaceManager) {
-        self.workspace = workspace
-        super.init(nibName: nil, bundle: nil)
-    }
-
-    required init?(coder: NSCoder) { fatalError() }
-
-    override func loadView() {
-        let container = ClearlySidebarBackgroundView()
-        container.wantsLayer = true
-        container.layer?.backgroundColor = Theme.sidebarBackground.cgColor
-        self.view = container
-
-        setupOutlineView()
-        setupEmptyState()
-        updateVisibility()
-
-        // Poll for workspace changes (simple and reliable)
-        updateTimer = Timer.scheduledTimer(withTimeInterval: 0.25, repeats: true) { [weak self] _ in
-            Task { @MainActor in
-                self?.outlineCoordinator?.reloadIfNeeded()
-                self?.updateVisibility()
-            }
-        }
-
-        // Observe fullscreen transitions to adjust top inset
-        let nc = NotificationCenter.default
-        fullscreenObservers.append(nc.addObserver(forName: NSWindow.didEnterFullScreenNotification, object: nil, queue: .main) { [weak self] _ in
-            Task { @MainActor in self?.updateTopInset(fullscreen: true) }
-        })
-        fullscreenObservers.append(nc.addObserver(forName: NSWindow.didExitFullScreenNotification, object: nil, queue: .main) { [weak self] _ in
-            Task { @MainActor in self?.updateTopInset(fullscreen: false) }
-        })
-    }
-
-    private func updateTopInset(fullscreen: Bool) {
-        let top: CGFloat = fullscreen ? 8 : 36
-        scrollView?.contentInsets.top = top
-        headerHeightConstraint?.constant = top
-    }
-
-    private func setupOutlineView() {
-        let coordinator = FileExplorerOutlineView.Coordinator(workspace: workspace)
-        self.outlineCoordinator = coordinator
-
-        let sv = NSScrollView()
-        sv.hasVerticalScroller = true
-        sv.autohidesScrollers = true
-        sv.drawsBackground = false
-        sv.translatesAutoresizingMaskIntoConstraints = false
-
-        let outlineView = FlatSectionOutlineView()
-        outlineView.headerView = nil
-        outlineView.style = .plain
-        outlineView.indentationPerLevel = 10
-        outlineView.rowSizeStyle = .custom
-        outlineView.rowHeight = 28
-        outlineView.selectionHighlightStyle = .regular
-        outlineView.floatsGroupRows = false
-        outlineView.intercellSpacing = NSSize(width: 0, height: 0)
-        outlineView.backgroundColor = .clear
-        outlineView.autoresizesOutlineColumn = false
-
-        let column = NSTableColumn(identifier: NSUserInterfaceItemIdentifier("name"))
-        column.isEditable = false
-        outlineView.addTableColumn(column)
-        outlineView.outlineTableColumn = column
-
-        outlineView.dataSource = coordinator
-        outlineView.delegate = coordinator
-
-        outlineView.autosaveName = "ClearlySidebarOutline"
-        outlineView.autosaveExpandedItems = true
-
-        let menu = NSMenu()
-        menu.delegate = coordinator
-        outlineView.menu = menu
-        outlineView.doubleAction = nil
-
-        // Drag and drop
-        outlineView.registerForDraggedTypes([.fileURL])
-        outlineView.setDraggingSourceOperationMask(.move, forLocal: true)
-
-        sv.documentView = outlineView
-        coordinator.outlineView = outlineView
-        outlineView.colorCoordinator = coordinator
-
-        // Make everything transparent so the container background shows through
-        sv.contentView.drawsBackground = false
-        sv.contentView.wantsLayer = true
-        sv.contentView.layer?.backgroundColor = NSColor.clear.cgColor
-
-        // Add content insets: top for traffic lights, left/right for breathing room
-        sv.contentInsets = NSEdgeInsets(top: 36, left: 6, bottom: 0, right: 6)
-        sv.scrollerInsets = NSEdgeInsets(top: 0, left: 0, bottom: 0, right: -6)
-        sv.automaticallyAdjustsContentInsets = false
-
-        view.addSubview(sv)
-        self.scrollView = sv
-
-        NSLayoutConstraint.activate([
-            sv.leadingAnchor.constraint(equalTo: view.leadingAnchor),
-            sv.trailingAnchor.constraint(equalTo: view.trailingAnchor),
-            sv.topAnchor.constraint(equalTo: view.topAnchor),
-            sv.bottomAnchor.constraint(equalTo: view.bottomAnchor),
-        ])
-
-        // Opaque header that covers the traffic lights area so content doesn't scroll behind it
-        let headerCover = ClearlySidebarBackgroundView()
-        headerCover.wantsLayer = true
-        headerCover.layer?.backgroundColor = Theme.sidebarBackground.cgColor
-        headerCover.translatesAutoresizingMaskIntoConstraints = false
-        view.addSubview(headerCover, positioned: .above, relativeTo: sv)
-        let heightConstraint = headerCover.heightAnchor.constraint(equalToConstant: 36)
-        self.headerHeightConstraint = heightConstraint
-        NSLayoutConstraint.activate([
-            headerCover.leadingAnchor.constraint(equalTo: view.leadingAnchor),
-            headerCover.trailingAnchor.constraint(equalTo: view.trailingAnchor),
-            headerCover.topAnchor.constraint(equalTo: view.topAnchor),
-            heightConstraint,
-        ])
-
-        DispatchQueue.main.async {
-            coordinator.reloadAndExpand()
-        }
-    }
-
-    private func setupEmptyState() {
-        let hostingView = NSHostingView(rootView: FileExplorerEmptyView(workspace: workspace))
-        hostingView.translatesAutoresizingMaskIntoConstraints = false
-        // Force transparent so container background shows through
-        hostingView.wantsLayer = true
-        hostingView.layer?.backgroundColor = NSColor.clear.cgColor
-        view.addSubview(hostingView)
-        self.emptyHostingView = hostingView
-
-        NSLayoutConstraint.activate([
-            hostingView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
-            hostingView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
-            hostingView.topAnchor.constraint(equalTo: view.topAnchor),
-            hostingView.bottomAnchor.constraint(equalTo: view.bottomAnchor),
-        ])
-    }
-
-    private func updateVisibility() {
-        let isEmpty = workspace.locations.isEmpty && workspace.recentFiles.isEmpty && workspace.openDocuments.isEmpty
-        scrollView?.isHidden = isEmpty
-        emptyHostingView?.isHidden = !isEmpty
-    }
-
-    deinit {
-        updateTimer?.invalidate()
-        fullscreenObservers.forEach { NotificationCenter.default.removeObserver($0) }
-    }
-}
-
-/// NSView subclass that maintains sidebar background across appearance changes
-class ClearlySidebarBackgroundView: NSView {
-    override func updateLayer() {
-        layer?.backgroundColor = Theme.sidebarBackground.cgColor
-    }
-
-    override var wantsUpdateLayer: Bool { true }
-
-    override func viewDidChangeEffectiveAppearance() {
-        super.viewDidChangeEffectiveAppearance()
-        needsDisplay = true
-        layer?.backgroundColor = Theme.sidebarBackground.cgColor
-    }
-}
